% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM_create_devs_list.R
\name{add_dev_changes}
\alias{add_dev_changes}
\title{Add the deviation changes from the list obj to an existing df}
\usage{
add_dev_changes(fut_list, scen, iter, par, dat, vals_df, nyrs)
}
\arguments{
\item{fut_list}{A single change input}

\item{scen}{The scenario name}

\item{iter}{The iteration name}

\item{par}{the parameter list}

\item{dat}{The data list}

\item{vals_df}{The dataframe with future om values}

\item{nyrs}{The number of years to extend the model forward}
}
\value{
A modified version of vals_df with the new changes applied.
}
\description{
Add the deviation changes from the list obj to an existing df
}
\author{
Kathryn Doering
}
