% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendOM.R
\name{add_sample_struct}
\alias{add_sample_struct}
\title{Add in years of sampling data needed}
\usage{
add_sample_struct(sample_struct, dat, nyrs_extend)
}
\arguments{
\item{sample_struct}{The sampling structure, as specified by the user.}

\item{dat}{A datafile as read in by r4ss::SS_readdat}

\item{nyrs_extend}{Number of years to extend the OM forward. Use 0 if using
this function on the initial (historical) run of the OM.}
}
\description{
Add in years of sampling data needed
}
