% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvirEM.R
\name{add_sample_struct_FixedCatches}
\alias{add_sample_struct_FixedCatches}
\title{Add FixedCatch and/or FixedCatchEM to sample_struct}
\usage{
add_sample_struct_FixedCatches(
  sample_struct,
  om_on = TRUE,
  em_on = TRUE,
  rt_year_om = c(2018, 2021),
  rt_year_em = c(2018, 2021),
  rt_fleet = 5,
  rt_mortality_om = 0.1,
  rt_mortality_em = 0.1,
  em_fixed = 0
)
}
\arguments{
\item{sample_struct}{The sample_struct object you want to add FixedCatch or FixedCatchEM to.}

\item{om_on}{TRUE if FixedCatch is on.}

\item{em_on}{TRUE if FixedCatchEM is on.}

\item{rt_year_om}{Years where the environmental event occur in the OM, "rt" stands for red tide since it was the original use case.}

\item{rt_year_em}{Years where the environmental event occur in the EM.}

\item{rt_fleet}{The number associated with the fleet that is being altered.}

\item{rt_mortality_om}{The mortality magnitude of the event in the OM, this is always fixed.}

\item{rt_mortality_em}{The mortality magnitude of the event in the EM, this is a placeholder unless the EM is fixed.}

\item{em_fixed}{If 0 the mortality in the EM will be estimated.  If 1 the mortality will be fixed to the provided rt_mortality_em in the EM.}
}
\value{
A sample_struct list object with appended FixedCatch and FixedCatchEM information.
}
\description{
After creating a base sample_struct, this function can add FixedCatch or
FixedCatchEM when scenarios require specified fixed catches in the OM or EM.
}
\examples{
# Create a sample_struct where the OM has a mortality event every 3 years, and the EM has one in every year.  
sample_struct_rep_3_x_all_yrs <- add_sample_struct_FixedCatches(sample_struct, rt_year_om = seq(from = 2018, to = 2047, by = 3), rt_year_em = seq(from = 2018, to = 2047, by = 1))
print(head(sample_struct_rep_3_x_all_yrs$FixedCatch))
print(head(sample_struct_rep_3_x_all_yrs$FixedCatchEM))

}
\author{
Kathryn Doering modified by Cassidy Peterson and Alexandra Norelli
}
