% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_error.R
\name{build_impl_error}
\alias{build_impl_error}
\title{Build a list of implementation error vectors for each scenario and iteration.}
\usage{
build_impl_error(
  yrs,
  nyrs_assess,
  n_impl_error_groups,
  scope,
  impl_error_pattern,
  impl_error_pars,
  scen_name_vec,
  iter_list
)
}
\arguments{
\item{yrs}{the number of years to simulate recruitment deviations for.}

\item{scope}{scope over which recruitment devations will be randomized.}

\item{scen_name_vec}{a vector of scenario names}

\item{iter_list}{a list of iteration names.}

\item{rec_dev_pattern}{how to simulate recruitment devations.}

\item{rec_dev_pars}{recruitment devation simulation parameters dependent on chosen pattern.}

\item{stddev}{the standard deviation of simulated recruitment deviations}
}
\value{
A list of scenarios with lists of interations in each with a vector of 
 rec devs for each simulation year.
}
\description{
Build a list of implementation error vectors for each scenario and iteration.
}
