% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdevs.R
\name{build_rec_devs}
\alias{build_rec_devs}
\title{Build an array of recruitment deviation vectors for every scenario and iteration}
\usage{
build_rec_devs(
  yrs,
  nyrs_assess,
  scope,
  rec_dev_pattern,
  rec_dev_pars,
  stddev,
  scen_name_length,
  iter_list
)
}
\arguments{
\item{yrs}{the number of years to simulate recruitment deviations for.}

\item{scope}{scope over which recruitment devations will be randomized.}

\item{rec_dev_pattern}{how to simulate recruitment devations.}

\item{rec_dev_pars}{recruitment devation simulation parameters dependent on chosen pattern.}

\item{stddev}{the standard deviation of simulated recruitment deviations}

\item{iter_list}{a list of iteration names.}

\item{scen_name_vec}{a vector of scenario names}
}
\value{
A list of scenarios with lists of interations in each with a vector of 
 rec devs for each simulation year.
}
\description{
Build an array of recruitment deviation vectors for every scenario and iteration
}
