% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_noise_funcs.R
\name{calc_comp_var}
\alias{calc_comp_var}
\title{Calculate uncertainty and biases in historic composition data. Not currently 
Used, but could be used in the future.}
\usage{
calc_comp_var(
  data_obs,
  data_exp,
  bins,
  fleets = NULL,
  years = NULL,
  seasons = NULL,
  merge_genders = TRUE,
  genders = NULL,
  merge_seasons = TRUE,
  merge_fleets = FALSE
)
}
\arguments{
\item{data_obs}{A data frame of observed composition data extracted from SS .dat file}

\item{data_exp}{A data frame of the expected composition data as estimated by an SS assessment model}

\item{bins}{A vector object including the composition bins}

\item{fleets}{A vector of the fleet numbers to analyze composition uncertainty for (Default is all fleets if NULL)}

\item{years}{A vector of the years to include when calculating composition uncertainty (Default is all years if NULL)}

\item{seasons}{A vector of the seasons to include when calculating composition uncertainty (Default is all years if NULL)}

\item{merge_genders}{TRUE/FALSE should genders be merged to calculate variance and biases (Defaults to TRUE)}

\item{genders}{A vector of the genders to analyze composition uncertainty for (Default is all genders if NULL)}

\item{merge_seasons}{TRUE/FALSE should seasons be merged to calculate variance and biases (Defaults to TRUE)}

\item{merge_fleets}{TRUE/FALSE should fleets be merged to calculate variance and biases (Defaults to FALSE)}
}
\value{
A list object with uncertainty and bias characteristics to inform data simulation.
}
\description{
Calculate uncertainty and biases in historic composition data. Not currently 
Used, but could be used in the future.
}
\author{
Nathan R. Vaughan
}
