% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EM.R
\name{change_data}
\alias{change_data}
\title{Change dataset from OM into format for EM}
\usage{
change_data(OM_datafile, EM_dir, do_checks = TRUE, verbose = FALSE)
}
\arguments{
\item{OM_datafile}{Relative or absolute path with filename to the
datafile to use.}

\item{EM_dir}{Absolute or relative path to the Estimation model directory.}

\item{do_checks}{Should checks on the data be performed? Defaults to TRUE.}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\description{
Change dataset from OM into format for EM
}
\examples{
\dontrun{
#TODO: Add example
}
}
