% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EM.R
\name{change_yrs_fcast}
\alias{change_yrs_fcast}
\title{Change the years in the forecast file}
\usage{
change_yrs_fcast(
  fore,
  make_yrs_rel = TRUE,
  nyrs_increment = NULL,
  nyrs_fore = NULL,
  mod_styr,
  mod_endyr
)
}
\arguments{
\item{fore}{A forecasting file read into R using r4ss::SS_readforecast()}

\item{make_yrs_rel}{Should the absolute years in the forecast file be changed
to relative years? Defaults to TRUE.}

\item{nyrs_increment}{The number of years to increment forecasting period years. 
If NULL (the default value), will not be incremented.}

\item{nyrs_fore}{The number of years of forecasting to do. If NULL, do not
change the number of forecasting years already specified in \code{fore}}

\item{mod_styr}{The first year of the model}

\item{mod_endyr}{The last year of the model \code{fore} assumes when read in.
Note that the assumed model year will be different for the output if
nyrs_increment is not NULL.}
}
\value{
A forecasting file as an R list object
}
\description{
This is both to increment years forward and/or to change absolute years to
relative years.
}
\author{
Kathryn Doering
}
