% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_avail_dat}
\alias{check_avail_dat}
\title{check all index years/fleets in EM available in OM. (but not vice versa)
a general function that can be used}
\usage{
check_avail_dat(
  EM_dat,
  OM_dat,
  list_item = "CPUE",
  colnames = c("year", "seas", "index")
)
}
\arguments{
\item{EM_dat}{An SS data file read in using r4ss for an EM}

\item{OM_dat}{An SS data file read in using r4ss for an OM}

\item{list_item}{A component in both EM_dat and OM_dat to check values for.
This should be a single string value.}

\item{colnames}{The column names of data to append together.}
}
\description{
check all index years/fleets in EM available in OM. (but not vice versa)
a general function that can be used
}
\author{
Kathryn Doering
}
