% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_catch_df}
\alias{check_catch_df}
\title{Check the catch dataframe}
\usage{
check_catch_df(df)
}
\arguments{
\item{df}{The catch datafarme to test}
}
\description{
Check the catch dataframe
}
\author{
Kathryn Doering
}
