% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_perf_metrics.R
\name{check_convergence}
\alias{check_convergence}
\title{Flag potential convergence issues in SS3 model runs}
\usage{
check_convergence(summary, min_yr, max_yr)
}
\arguments{
\item{summary}{Summary returned from running `SSMSE_summary_all()`}

\item{min_yr}{The first year of SSB checked}

\item{max_yr}{The last year of SSB checked}
}
\value{
A tibble containing the SSB values in the EM relative to the OM by
 model run of each iteration of each scenario.
}
\description{
Does basic checks for convergance of estimation model runs from `run_SSMSE()`
 simulations. This function 1) warns if there are parameters on bounds; 2)
 warns if the SSB in the EM is 2x as large or half as small as the OM. Note
 these warnings may not mean that the models have not converged, but can flag
 potential issues that can be investigated further
}
\examples{
\dontrun{
check_convergance(SSMSE_summary, min_yr = 101, max_yr = 120)
}
}
