% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_dat_str}
\alias{check_dat_str}
\title{Check dat_str_list}
\usage{
check_dat_str(dat_str)
}
\arguments{
\item{dat_str}{The list to check. Should be a list including which years and
fleets should be added from the OM into the EM for different types of data.}
}
\description{
Check that list object dat_str_list has the expected form, including the
correct names, correct column names (as in r4ss), and that all values in the
dataframes are integer or numeric. This does not check for if numeric or
interger values make sense given the model used.
}
\author{
Kathryn Doering
}
