% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_dir}
\alias{check_dir}
\title{Check that the directory for an OM is valid}
\usage{
check_dir(dir)
}
\arguments{
\item{dir}{Input to check. Should be a directory name that should contain an
SS3 model that can be used as an OM.
@author Kathryn Doering}
}
\description{
Check that the directory contains starter and forecast SS3 files.
}
