% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_EM_forecast}
\alias{check_EM_forecast}
\title{Check structure of forecast is suitable to use in the EM}
\usage{
check_EM_forecast(fore, n_flts_catch = NULL)
}
\arguments{
\item{fore}{A forecast list read in using r4ss::SS_readforecast}

\item{n_flts_catch}{The number of fleets with catch. If NULL, this function
will skip a check requiring this input.}
}
\value{
Function mainly used for side effects, but returns TRUE invisibly if
 no errors created.
}
\description{
Check structure of forecast is suitable to use in the EM
}
\author{
Kathryn Doering
}
