% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_om.R
\name{check_future_om_list_str}
\alias{check_future_om_list_str}
\title{Check the general structure of a future OM list and standardize values}
\usage{
check_future_om_list_str(future_om_list)
}
\arguments{
\item{future_om_list}{The future_om_list object to check}
}
\value{
The future_om_list with implicit arguments made explicit
}
\description{
Checks that a future OM list is valid. If any values are implicit, 
then add these values. Does not check against arguments in the scenario,
just the generic structure
}
