% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_error.R
\name{check_impl_error}
\alias{check_impl_error}
\title{check that user input implementation error vectors are the correct length}
\usage{
check_impl_error(length_impl_error_seq, expected_length)
}
\arguments{
\item{length_impl_error_seq}{The length of the input implementation error sequence}

\item{expected_length}{the expected length of the implementation error sequence}
}
\description{
check that user input implementation error vectors are the correct length
}
