% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_OM_dat}
\alias{check_OM_dat}
\title{check that an OM data set has at least the same data as an estimation model}
\usage{
check_OM_dat(OM_dat, EM_dat)
}
\arguments{
\item{OM_dat}{A data set read in using r4ss::SS_readdat from an operating 
model. Note that it should span the same years as EM_dat.}

\item{EM_dat}{A data set read in using r4ss::SS_readdata from an estimation 
model. Note that it should span the same years as EM_dat}
}
\description{
check that an OM data set has at least the same data as an estimation model
}
