% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdevs.R
\name{check_recdev_error}
\alias{check_recdev_error}
\title{check that user input recruitment deviation vectors are the correct length}
\usage{
check_recdev_error(length_recdev_seq, expected_length)
}
\arguments{
\item{length_recdev_seq}{The length of the input recruitment deviation sequence}

\item{expected_length}{the expected length of the recruitment deviation sequence}
}
\description{
check that user input recruitment deviation vectors are the correct length
}
