% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_scen_list}
\alias{check_scen_list}
\title{Check structure of the object scen_list}
\usage{
check_scen_list(list, verbose = FALSE)
}
\arguments{
\item{list}{A list to check}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\description{
Check the structure that is input to \code{\link{run_SSMSE}}.
}
