% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_cols}
\alias{combine_cols}
\title{function that creates a combined column to the list_item of interest}
\usage{
combine_cols(dat_list, list_item, colnames)
}
\arguments{
\item{dat_list}{An SS3 data file as a list read in using r4ss}

\item{list_item}{List item in dat_list to extract and return a modified
version of this value}

\item{colnames}{Column names in list_item}
}
\description{
function that creates a combined column to the list_item of interest
}
