% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_om.R
\name{create_future_om_list}
\alias{create_future_om_list}
\title{Helper function to create future om list objects}
\usage{
create_future_om_list(
  example_type = c("model_change", "custom"),
  list_length = 1
)
}
\arguments{
\item{example_type}{Type of example future_om_list object to create. Options are
"model_change" or "custom". Defaults to "model_change".}

\item{list_length}{The length of the example list to create. Defaults to 1.
For now, just replicates the same list.}
}
\description{
The future_om_list objects specify changes to make in the future to the
OM as the OM is extended forward in time. In particular, this function helps
users create these objects. For now, just returns examples based on cod
model that comes with SSMSE.
}
\examples{
example_future_om_list <-
  create_future_om_list(example_type = "custom", list_length = 2)
}
