% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_out_dirs}
\alias{create_out_dirs}
\title{create the OM directory}
\usage{
create_out_dirs(
  out_dir,
  niter,
  OM_name,
  OM_in_dir,
  EM_name = NULL,
  EM_in_dir = NULL
)
}
\arguments{
\item{out_dir}{The directory to which to write output. IF NULL, will default
to the working directory.}

\item{niter}{The number iteration}

\item{OM_name}{Name of the operating model (OM).}

\item{OM_in_dir}{Relative or absolute path to the operating model, if using a
model outside of the SSMSE package. Should be a string.}

\item{EM_name}{Name of the EM model}

\item{EM_in_dir}{Relative or absolute path to the estimation model,}
}
\value{
A list with 2 named components each of length 1 characters. The
 components are: OM_dir, where OM will be run, and OM_in_dir, where the model
 files will be copied from.
}
\description{
Create an OM directory within the out_dir specified (named by the value of
niter)
}
