% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_struct.R
\name{create_sample_struct}
\alias{create_sample_struct}
\title{Create the sample_struct list}
\usage{
create_sample_struct(dat, nyrs)
}
\arguments{
\item{dat}{An r4ss list object read in using r4ss::SS_readdat() or the path
(relative or absolute) to an SS data file to read in.}

\item{nyrs}{Number of years beyond the years included in the dat file to run
the MSE. A single integer value.}
}
\description{
Create a sampling strcture list using the pattern in a data file and a year
range.
}
\examples{
OM_path <- system.file("extdata", "models", "cod", "ss3.dat", package = "SSMSE")
# note there is a warning for lencomp because it does not have a consistent pattern
sample_struct <- create_sample_struct(OM_path, nyrs = 20)
print(sample_struct)
}
\author{
Kathryn Doering
}
