% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatioBiasEM.R
\name{create_sample_struct_hist}
\alias{create_sample_struct_hist}
\title{Create the sample_struct list for historical data}
\usage{
create_sample_struct_hist(dat, rm_NAs = FALSE)
}
\description{
Create the sample_struct list for historical data
}
