% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvirEM.R
\name{create_sample_struct_hist_envir}
\alias{create_sample_struct_hist_envir}
\title{Create the sample_struct list for historical data}
\usage{
create_sample_struct_hist_envir(dat, rm_NAs = FALSE)
}
\description{
Create the sample_struct list for historical data
}
