% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop_OMs.R
\name{develop_OMs}
\alias{develop_OMs}
\title{Develop different operating models}
\usage{
develop_OMs(
  OM_name = NULL,
  OM_in_dir = NULL,
  out_dir = getwd(),
  par_name,
  par_vals,
  refit_OMs = TRUE,
  hess = FALSE
)
}
\arguments{
\item{OM_name}{Name of the original model, as in the SSMSE package. If not
using a model in the package, please specify its path in \code{OM_in_dir}}

\item{OM_in_dir}{Path to the original operating model. If using a model in
the SSMSE package, please specify its name in \code{OM_name} instead.}

\item{out_dir}{Path where the new models will be written. Defaults to the
current working directory.}

\item{par_name}{Name of the parameter to modify}

\item{par_vals}{Vector of parameter values to modify in the OM. Assume these
will be fixed so phase will be set as negative.}

\item{refit_OMs}{Should the models be refit to data? Defaults to TRUE}

\item{hess}{Should the hessian be estimated if reffiting the OMs? defaults to
FALSE}
}
\description{
THis is a utility to help a user create new operating models starting from the
same model. For now, it is only possible to adjust 1 parameter value
}
