% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_functions.R
\name{EM}
\alias{EM}
\title{Use EM as the management strategy option.}
\usage{
EM(
  EM_out_dir = NULL,
  init_loop = TRUE,
  OM_dat,
  verbose = FALSE,
  nyrs_assess,
  dat_yrs,
  sample_struct = NULL,
  sample_struct_hist = NULL,
  seed = NULL,
  OM_out_dir,
  ...
)
}
\arguments{
\item{EM_out_dir}{Relative or absolute path to the estimation model, if using a
model outside of the SSMSE package.}

\item{init_loop}{Logical. If this is the first initialization loop of the
MSE, \code{init_loop} should be TRUE. If it is in further loops, it should
be FALSE.}

\item{OM_dat}{An valid SS data file read in using r4ss. In particular,
this should be sampled data.}

\item{verbose}{Want verbose output? Defaults to FALSE.}

\item{nyrs_assess}{The number of years between assessments. E.g., if an
assessment is conducted every 3 years, put 3 here. A single integer value.}

\item{dat_yrs}{Which years should be added to the new model? Ignored if
init_loop is TRUE.}

\item{sample_struct}{A optional list including which years, seasons, and fleets
should be  added from the OM into the EM for different types of data.
If NULL, the data structure will try to be inferred from the pattern found
for each of the datatypes within the EM datafiles. Include this structure
for the number of years to extend the model out. Note that the data should
be specified using the list component names and column names as in would be
used in \code{r4ss::SS_readdat()}. The \code{run_SSMSE_iter} function
examples give an example of what this structure should be. Running the
function \code{create_sample_struct()} will also produce a sample_struct
object in the correct form. Can be NULL only when MS is not EM.}

\item{sample_struct_hist}{historical sample structure object}

\item{seed}{A random seed so that reproducible results are possible.}

\item{OM_out_dir}{The full path to the directory in which the OM is run.}

\item{...}{Any additional parameters}
}
\description{
Use EM as the management strategy option.
}
\author{
Kathryn Doering
}
