% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MS.R
\name{extend_EM_bias_adj}
\alias{extend_EM_bias_adj}
\title{Extend the EM bias adjustment forward in time}
\usage{
extend_EM_bias_adj(ctlfile, datlist_new, nyrs_assess, write_ctl)
}
\arguments{
\item{ctlfile}{Path to the control file.}

\item{datlist_new}{A data file read in usint r4ss::SS_readdat that has already
been extended forward}

\item{write_ctl}{Should the new controlfile be written, overwritting the old
one?}
}
\value{
The new control file with recdev values extended forward by
 nyrs_assess
}
\description{
Extends the EN bias adjustment forward the same amount as the number of
assessment years. A simple default way of doing this.
}
