% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendOM.R
\name{extend_OM}
\alias{extend_OM}
\title{Extend the OM forward using next years' catch}
\usage{
extend_OM(
  catch,
  OM_dir,
  dummy_dat_scheme = NULL,
  nyrs_extend = 3,
  write_dat = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{catch}{A dataframe of catch values and its associated information to
 add to the OM. The column names are the same as in an SS data file (e.g., 
 year,    season, fleet,    catch,    catch_se).
length of the number of years (only works when catch is for 1 fleet)}

\item{OM_dir}{The full path to the OM directory.}

\item{dummy_dat_scheme}{The sampling scheme for dummy data. Current options
are NULL, which means no new dummy data will be added CPUE, length comps, or
age comps, or \code{"all"}, which means that new years of data will be added 
for CPUE, length comps (if using), and age comps, for all new years and 
combinations of fleet and season. For length comps and age comps, a dummy
value will also be added for each combination of other metadata values,
including (but not limited to ) sex, partition, and bin high and low values.}

\item{nyrs_extend}{An integer value of years to extend the model forward. Defaults
to an arbitrary value of 3.}

\item{write_dat}{Should the datafile be overwritten? Defaults to TRUE.}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\value{
A new dat list object (format as created by r4ss::SS_readdat) that
 has been extended forward  as if read in by r4ss function SS_readdat
}
\description{
Add in the catch values for the next years, extend the OM forward for the 
appropriate number of years.
}
