% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_extend_OM.R
\name{extend_OM_full_period}
\alias{extend_OM_full_period}
\title{Add in future parameter values}
\usage{
extend_OM_full_period(
  ctl,
  dat,
  parlist,
  future_om_list,
  nyrs,
  nscen,
  scen_name,
  niter
)
}
\arguments{
\item{dat}{A datafile as read in by r4ss::SS_readdat}

\item{parlist}{A parameter file as read in by r4ss::SS_readpar_3.30}

\item{future_om_list}{A future OM parameter specification object}

\item{nyrs}{The total number of years to run the MSE}

\item{nscen}{The scenario number}

\item{scen_name}{The scenario name}

\item{niter}{The iteration number}

\item{clt}{A control file as read in by r4ss::SS_readctl}
}
\description{
Add in future parameter values
}
