% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_perf_metrics.R
\name{get_avg_catch}
\alias{get_avg_catch}
\title{Example Performance Metric: Calculate average catch over a range of years}
\usage{
get_avg_catch(datfile, yrs)
}
\arguments{
\item{datfile}{Path to the Stock Synthesis data file containing catch}

\item{yrs}{A vector containing a range of years. Years are as defined in the
Stock Synthesis data file.}
}
\value{
The average catch, a number.
}
\description{
Example performance metric that calculates average catch over a range of
 years in a Stock Synthesis data file. This function aggregates across
 fleets, so may not be appropriate for models with multiple fleets.
}
\examples{
\dontrun{
avg_catch <- function(datfile = "ss3model/dat.ss", yrs = 30:75) {
  avg_catch
}
}
}
