% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSS.R
\name{get_bin}
\alias{get_bin}
\title{Get SS3 binary/executable location in package}
\usage{
get_bin(bin_name = "ss3")
}
\arguments{
\item{bin_name}{Name of SS3 binary, defaults to "ss3"}
}
\value{
The path to an SS3 binary. If using the GitHub version of the
package, this will be an internal binary. Otherwise, this function
will search for a version of the binary in your path. See the
ss3sim vignette.
}
\description{
Get the binary/executable location in the package SSMSE. This function
is from \href{https://github.com/ss3sim/ss3sim}{ss3sim}.
}
\examples{
\dontrun{
get_bin()
}
}
