% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_F.R
\name{get_catch_retained}
\alias{get_catch_retained}
\title{Get retained catch from the timeseries Report.sso table}
\usage{
get_catch_retained(timeseries, units_of_catch)
}
\arguments{
\item{timeseries}{from SSoutput}

\item{units_of_catch}{From datalist}
}
\value{
a data frame with retained catch by Area, Yr, Era, Seas, Fleet, and 
 units (long format)
}
\description{
Get retained catch from the timeseries Report.sso table
}
