% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_extend_OM.R
\name{get_dev_loc}
\alias{get_dev_loc}
\title{Add in future parameter values}
\usage{
get_dev_loc(ctl)
}
\arguments{
\item{clt}{A control file as read in by r4ss::SS_readctl}
}
\description{
Add in future parameter values
}
