% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendOM.R
\name{get_dummy_dat}
\alias{get_dummy_dat}
\title{Get the dummy data for a data type}
\usage{
get_dummy_dat(dummy_dat_scheme)
}
\arguments{
\item{dummy_dat_scheme}{The years, fleets, data types etc. to add}
}
\description{
Get the dummy data for a data type
}
