% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_functions.R
\name{get_EM_catch_df}
\alias{get_EM_catch_df}
\title{Get the EM catch data frame}
\usage{
get_EM_catch_df(EM_dir, dat)
}
\arguments{
\item{EM_dir}{Path to the EM files}

\item{dat}{A SS3 datfile read into R using \code{r4ss::SS_readdat()}}
}
\value{
A data frame of future catch
}
\description{
Get the data frame of catch for the next iterations when using a Stock
Synthesis Estimation model from the Report.sso file.
}
\author{
Kathryn Doering
}
