% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EM.R
\name{get_EM_dat}
\alias{get_EM_dat}
\title{Change the OM data to match the format of the original EM data}
\usage{
get_EM_dat(OM_dat, EM_dat, do_checks = TRUE)
}
\arguments{
\item{OM_dat}{An SS3 data file read in by as a list read in using r4ss from
the operating model}

\item{EM_dat}{An SS3 data file read in by as a list read in using r4ss from
the estimation model}

\item{do_checks}{Should checks on the data be performed? Defaults to TRUE.}
}
\value{
A data list in the same format that can be read/written by r4ss that
has index, lcomps, and age comps from OM_dat, but with the same structure as
EM_dat.
}
\description{
This does the technical part of changing the EM data. Note this may be unnecessary
}
\author{
Kathryn Doering
}
