% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_struct.R
\name{get_full_sample_struct}
\alias{get_full_sample_struct}
\title{Get the full sample structure from user input}
\usage{
get_full_sample_struct(sample_struct, OM_out_dir)
}
\arguments{
\item{sample_struct}{A optional list including which years, seasons, and fleets
should be  added from the OM into the EM for different types of data.
If NULL, the data structure will try to be inferred from the pattern found
for each of the datatypes within the EM datafiles. Include this structure
for the number of years to extend the model out. Note that the data should
be specified using the list component names and column names as in would be
used in \code{r4ss::SS_readdat()}. The \code{run_SSMSE_iter} function
examples give an example of what this structure should be. Running the
function \code{create_sample_struct()} will also produce a sample_struct
object in the correct form. Can be NULL only when MS is not EM.}

\item{OM_out_dir}{The full path to the directory in which the OM is run.}
}
\value{
A list of the full sample structure, using names as input by the user
input by the user (not r4ss names).
}
\description{
Get the full sample structure from user input by looking at the OM data. If it
cannot be unambiguously determined, this function will return an error
describing what additional user input is required.
}
