% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{get_init_samp_scheme}
\alias{get_init_samp_scheme}
\title{Get the sampling scheme in a data file.}
\usage{
get_init_samp_scheme(
  dat,
  dat_types = c("CPUE", "lencomp", "agecomp", "meanbodywt", "MeanSize_at_Age_obs")
)
}
\arguments{
\item{dat}{An SS3 data file}

\item{dat_types}{Types of data to include}
}
\description{
Determine what the default sampling scheme is for a given data file.
Produces a list object with the sampling scheme, which can be modified, if
desired.
}
