% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdevs.R
\name{get_inp_mean_recs}
\alias{get_inp_mean_recs}
\title{calculate input mean to achieve target mean}
\usage{
get_inp_mean_recs(target_mean, inp_var)
}
\arguments{
\item{target_mean}{The target mean of the observed log normal values}

\item{inp_var}{the input variance to the random log normal generator}
}
\value{
the input mean required to achieve target mean.
}
\description{
calculate input mean to achieve target mean
}
