% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_error.R
\name{get_inp_var}
\alias{get_inp_var}
\title{calculate the input variance from target mean and variance}
\usage{
get_inp_var(target_mean, targ_var)
}
\arguments{
\item{target_mean}{The target mean of the observed log normal values}

\item{targ_var}{the target variance to the simulated log normal values}
}
\value{
the squared difference between observed and target variances
}
\description{
calculate the input variance from target mean and variance
}
