% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_input_value}
\alias{get_input_value}
\title{return a value from a data frame}
\usage{
get_input_value(data, method = "most_common_value", colname, group = NULL)
}
\arguments{
\item{data}{A dataframe which has a column that matches (at least partially)
colname}

\item{method}{How should the value to be returned be selected? Current
options include "most_common_value", where the most common input uncertainty
value will be returned and "only_value" where all input values must be the
same in data; if they are, this value will be returned. Otherwise, an error
will be generated.}

\item{colname}{Column name as a string in \code{data}. Note that partial
matching and regular expressions can be used.}

\item{group}{Column name as a string in \code{data} used to group the data
before calculating the input value to use. Defaults to NULL.}
}
\value{
A value of the same type as \code{data[, colname]} if \code{group}
is NULL, or a data.frame if \code{group} is specified.
}
\description{
Return a single value from a column of a dataframe using the method specified
}
\details{
Note that this function was created initially to return a value to
use as the input uncertainty, but it should be generalizable to pulling a
value from a column in any data frame using the method specified.
}
\examples{
dfr <- data.frame(
  "year" = 1:5,
  "value" = c(2, 2, 2, 3, 3),
  "se_log" = 0.2
)
SSMSE:::get_input_value(
  data = dfr, method = "most_common_value", colname = "se_log",
  group = "value"
)
SSMSE:::get_input_value(data = dfr, method = "most_common_value", colname = "value")
SSMSE:::get_input_value(data = dfr, method = "only_value", colname = "se_log")
# generates an error:
# SSMSE:::get_input_value(data = dfr, method = "only_value", colname = "value")
}
\author{
Kathryn Doering
}
