% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatioBiasEM.R
\name{get_PercentEM_catch_df}
\alias{get_PercentEM_catch_df}
\title{Get the EM catch data frame}
\usage{
get_PercentEM_catch_df(
  EM_dir,
  dat,
  dat_yrs,
  nyrs_assess,
  changeup = 10,
  changedown = 10,
  c1 = 1,
  option = "default",
  ...
)
}
\arguments{
\item{EM_dir}{Path to the EM files}

\item{dat}{A SS datfile read into R using \code{r4ss::SS_readdat()}}
}
\value{
A data frame of future catch
}
\description{
Get the data frame of catch for the next iterations when using a Stock
Synthesis Estimation model from the Report.sso file.
}
\author{
Kathryn Doering modified by Cassidy Peterson
}
