% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{get_performance_metrics}
\alias{get_performance_metrics}
\title{get basic data to calculate performance metrics}
\usage{
get_performance_metrics(
  dir = getwd(),
  use_SSMSE_summary_all = TRUE,
  quantities = c("catch", "SpawnBio")
)
}
\arguments{
\item{dir}{Path to the directory containing the scenarios, either relative
or absolute. Defaults to the working directory.}

\item{use_SSMSE_summary_all}{If it exists, should the summmary files generated
by SSMSE_summary_all be used? Defaults to TRUE.}

\item{quantities}{Quantites from the operating model to add}
}
\description{
get basic data to calculate performance metrics
}
