% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_perf_metrics.R
\name{get_rel_SSB_avg}
\alias{get_rel_SSB_avg}
\title{Example Performance Metric: Calculate the avg relative SSB (SSB/SSB unfished)
over a range of years for each iteration}
\usage{
get_rel_SSB_avg(summary, min_yr, max_yr)
}
\arguments{
\item{summary}{Summary returned from running `SSMSE_summary_all()`}
}
\value{
A tibble containing the relative avg SSB per year by iteration and
 scenario.
}
\description{
Example performance metric that calculates the average Spawning Stock Biomass
 SSB (units as in the simulations) relative to the  unfished SSB over a range
 of years for each iteration of each scenario in the SSMSE simulation run.
}
\examples{
\dontrun{
rel_avg_ssb <- get_rel_SSB_avg(run_SSMSE_summary, min_yr = 10, max_yr = 105)
rel_avg_ssb
}
}
