% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_F.R
\name{get_retained_catch}
\alias{get_retained_catch}
\title{Get retained catch from the timeseries Report.sso table}
\usage{
get_retained_catch(timeseries, units_of_catch)
}
\arguments{
\item{timeseries}{The timeseries table from \code{r4ss::SS_output()}.}

\item{units_of_catch}{From datalist, the catch units. A named list where the
names are the fleets (to provide an extra check)}
}
\value{
a data frame with retained catch by Yr, Era, Seas, Fleet, and
units (long format)
}
\description{
Get retained catch from the timeseries Report.sso table
}
