% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_perf_metrics.R
\name{get_SSB_avg}
\alias{get_SSB_avg}
\title{Example Performance Metric: calculate the average SSB over a range of years
for each iteration}
\usage{
get_SSB_avg(summary, min_yr, max_yr)
}
\arguments{
\item{summary}{Summary returned from running \code{SSMSE_summary_all()}}

\item{min_year}{The first year to include in the average}

\item{max_year}{The last year to include in the average}
}
\value{
A tibble containing the average SSB by iteration and scenario.
}
\description{
Example performance metric that calculates the average Spawning Stock Biomass
SSB (units as in the simulations) over a range of years for each iteration of
each scenario in the SSMSE simulation run.
}
\examples{
\dontrun{
avg_ssb <- get_SSB_avg(run_SSMSE_summary, min_yr = 10, max_yr = 105)
avg_ssb
}
}
