% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM_create_devs_list.R
\name{match_parname}
\alias{match_parname}
\title{Match parameter name to parameter names in the par file}
\usage{
match_parname(list_pars, par)
}
\arguments{
\item{list_pars}{the parameter names to find}

\item{par}{the parfile in which to find list_pars}
}
\value{
A dataframe containing the parameter name and which object it is in
 in the par object.
}
\description{
Match parameter name to parameter names in the par file
}
\author{
Kathryn Doering
}
