% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_functions.R
\name{no_catch}
\alias{no_catch}
\title{No Catch in the future management strategy}
\usage{
no_catch(OM_out_dir, OM_dat, dat_yrs, ...)
}
\arguments{
\item{OM_out_dir}{The full path to the directory in which the OM is run.}

\item{OM_dat}{An valid SS3 data file read in using r4ss. In particular,
this should be sampled data.}

\item{dat_yrs}{Which years should be added to the new model? Ignored if
init_loop is TRUE.}

\item{...}{Any additional parameters}
}
\description{
No Catch in the future management strategy
}
\author{
Kathryn Doering
}
