% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{plot_comp_sampling}
\alias{plot_comp_sampling}
\title{Plot comp data, expected values, and sampled data for 1 scenario}
\usage{
plot_comp_sampling(dir = getwd(), comp_type = c("agecomp", "lencomp"))
}
\arguments{
\item{dir}{Path to the directory containing 1 scenario. Defaults to
the current working directory.}

\item{comp_type}{Type of composition data, age or length. Defaults to age.}
}
\value{
A list containing 2 components: 1) the ggplot object and 2) the
dataframe used to make the ggplot object
}
\description{
Creates a plot that can be used to see how sampling lines up with
data and expected values for the index of abundance
}
\author{
Kathryn Doering
}
