% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{r4ss_obj_err}
\alias{r4ss_obj_err}
\title{Error if object is not an r4ss object}
\usage{
r4ss_obj_err(obj_name = "object ", type = "list")
}
\arguments{
\item{obj_name}{Object name that is not an r4ss object to print in the error}

\item{type}{Type that obj_name was expected to be, but is not,}
}
\description{
Error if object is not an r4ss object
}
