% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{rm_sample_struct_hist}
\alias{rm_sample_struct_hist}
\title{Remove the historical sampling structure}
\usage{
rm_sample_struct_hist(sample_struct_hist, dat)
}
\arguments{
\item{sample_struct_hist}{An optional list including which years should be
sampled for the historical period for the data generated from the OM. If
this is left as NULL, then the same sampling scheme will be used as in the
OM's data file. If it is not NULL, then each year.}

\item{dat}{The data file, as read in using r4ss}
}
\description{
Remove the historical sampling structure
}
