% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{run_Base}
\alias{run_Base}
\title{Initial run of the Base model to get expected and bootstrap values}
\usage{
run_Base(OM_dir, verbose = FALSE, debug_par_run = TRUE, seed = NULL)
}
\arguments{
\item{OM_dir}{The full path to the OM directory.}

\item{verbose}{Want verbose output? Defaults to FALSE.}

\item{debug_par_run}{If set to TRUE, and the run fails, a new folder called
error_check will be created, and the model will be run from control start
values instead of ss.par. The 2 par files are then compared to help debug
the issue with the model run. Defaults to TRUE.}

\item{seed}{A random seed so that reproducible results are possible.}
}
\description{
This function is used to initialize the Base model and get expected values
for the OM and bootstraps for the EM.
}
\author{
Kathryn Doering
}
