% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EM.R
\name{run_EM}
\alias{run_EM}
\title{Run the estimation model}
\usage{
run_EM(
  EM_dir,
  hess = FALSE,
  check_converged = TRUE,
  nyrs_proj,
  set_use_par = FALSE,
  change_fcast = FALSE,
  seas = 1,
  catch_se = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{EM_dir}{Absolute or relative path to the estimation model directory}

\item{hess}{Get the hessian during model run? Defaults to FALSE. Not
estimating the hessian will speed up the run, but no estimates of error will
be generated.}

\item{check_converged}{Perform checks to see if the model converged? Defaults
to TRUE.}

\item{nyrs_proj}{The number of years of catch to be returned. Only used if
change_fcast = TRUE.}

\item{set_use_par}{Should input values be read from the .par file? If TRUE,
will change setting in the starter file; otherwise, will use the setting
already in the starter file, which may or may not read from the .par file.}

\item{change_fcast}{Should number of years be changed in forecast? Defaults
to TRUE.}

\item{seas}{Season for catch to be output. Defaults to 1.}

\item{catch_se}{Catch standard error for catch to be output. Defaults to 
0.01.}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\value{
The future catch in a dataframe, based on the EM run and the harvest
strategy outlined within the SS forecasting module.
}
\description{
Runs the estimation model and performs checks if desired.
}
