% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM_create_devs_list.R
\name{sample_vals}
\alias{sample_vals}
\title{Sample vals from normal random, lognormal random, or modified AR-1 
process.}
\usage{
sample_vals(
  mean,
  sd,
  ar_1_phi = 0,
  seed,
  ndevs,
  dist = c("normal", "lognormal")
)
}
\arguments{
\item{mean}{A single value or vector of mean parameters}

\item{sd}{A single value or vector of sd parameter}

\item{ar_1_phi}{The phi (coefficient) value for an ar 1 process. Should be 
between -1 and 1. 0 means an AR 1 process will NOT be used. 1 indicates a
random walk. A single value or vector.}

\item{seed}{The seed that will be set before sampling}

\item{ndevs}{The number of sampled values to expect}

\item{dist}{The distribution to sample from.}
}
\description{
Sample vals from normal random, lognormal random, or modified AR-1 
process.
}
\author{
Kathryn Doering
}
