% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_MSE_seeds}
\alias{set_MSE_seeds}
\title{Set the initial global, scenario, and iteration seeds}
\usage{
set_MSE_seeds(seed = NULL, iter_vec)
}
\arguments{
\item{seed}{Input a fixed seed to replicate previous simulation runs. Seed
can be a single value for a global seed, n_scenarios+1 length vector for
scenario specific and a global seed, n_iterations+n_scenarios+1 length
vector for iteration scenario and global seeds. Can also be a list object
with a single value under \code{seed[["global"]]}, a vector under
\code{seed[["scenario"]]}, and a multiple vectors for iteration specific seeds
under \code{seed[["iter"]][[1:n_scenarios]]}.}

\item{iter_vec}{The number of iterations per scenario. A vector of integers
in the same order as scen_name_vec.}
}
\value{
A list of length 3 with 1) the global seed value; 2) the scenario seed values; and 3) the iteration seed values.
}
\description{
Set the initial global, scenario, and iteration seeds
}
\examples{
seeds <- set_MSE_seeds(seed = seq(10, 80, by = 10), iter_vec = c(2, 3))
}
