% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/red_noise_funcs.R
\name{Sim_comp}
\alias{Sim_comp}
\title{Calculate uncertainty and biases in historic composition data}
\usage{
Sim_comp(
  Comp_uncert,
  data_exp,
  bins,
  years = NULL,
  seasons = NULL,
  fleets = NULL,
  genders = NULL
)
}
\arguments{
\item{Comp_uncert}{A list object representing the output from the calc_comp_var function}

\item{data_exp}{A vector representing the expected composition values for which to draw a random observation dataset}

\item{bins}{A vector object including the composition bins}

\item{years}{A vector of the years to simulate data for. default is all years in data_exp if NULL.}

\item{seasons}{A vector of the seasons to simulate data for. default is all seasons in data_exp  if NULL.}

\item{fleets}{A vector of the fleets to simulate data for. default is all fleets in data_exp  if NULL.}

\item{genders}{A vector of the genders to simulate data for. default is all genders in data_exp  if NULL.}
}
\value{
A list object with uncertainty and bias characteristics to inform data simulation.
}
\description{
Calculate uncertainty and biases in historic composition data
}
\author{
Nathan R. Vaughan
}
