% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{SSMSE_summary_iter}
\alias{SSMSE_summary_iter}
\title{Get results in a list for 1 iteration}
\usage{
SSMSE_summary_iter(dir)
}
\arguments{
\item{dir}{Path to the directory for 1 iteration of 1 scenario.}
}
\value{
A list of 3 data frames called scalar, timeseries, and
derived (for derived quantities). These lists contain information for
multiple model runs (estimation models and operating models) for 1
iteration.
}
\description{
Get results in a list for 1 iteration, using ss3sim::get_results_iter
}
\seealso{
\code{\link[ss3sim]{get_results_iter}}
}
