% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{SSMSE_summary_scen}
\alias{SSMSE_summary_scen}
\title{Get results in a list for 1 scenario}
\usage{
SSMSE_summary_scen(dir = getwd())
}
\arguments{
\item{dir}{Path to the directory for 1 scenario, either relative or absolute.
Defaults to the working directory.}
}
\value{
A list of 3 data frames called scalar, ts, and
dq (for derived quantities). These lists contain information for
multiple model runs (estimation models and operating models) for 1
iteration.Also writes 3 .csv files with the contents of this list of
dataframes to dir.
}
\description{
Get results in a list for 1 iteration, using ss3sim::get_results_iter
}
\seealso{
\code{\link[ss3sim]{get_results_scenario}}
}
