% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develtools.R
\name{test_no_par}
\alias{test_no_par}
\title{Change a model from running with par to running without par}
\usage{
test_no_par(orig_mod_dir, new_mod_dir)
}
\arguments{
\item{orig_mod_dir}{The original model directory}

\item{new_mod_dir}{The new model directory (folder need not exist)}
}
\description{
The intention of this function is to help troubleshooting issues with the
par file. It is intended mostly to help troubleshooting while developing
the SSMSE package, but may also be helpful with runtime testing.
}
