% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_extend_OM.R
\name{update_basevals_blocks}
\alias{update_basevals_blocks}
\title{Update a sequence of base parameter annual values to account for a time varying block effects}
\usage{
update_basevals_blocks(
  base_vals,
  base_years,
  temp_block,
  current_par,
  ctl,
  dat,
  temp_ctl,
  base_range,
  baseparm,
  base_bounds
)
}
\arguments{
\item{base_vals}{A vector of base parameter values that will be updated to include the impact of a time varying block change}

\item{base_years}{A vector of years for which the base values are needed}

\item{temp_block}{The time-varying parameter lines for the block effects on the base parameter}

\item{current_par}{The index of the current parameter being updated}

\item{ctl}{A control file as read in by \code{r4ss::SS_readctl}.}

\item{dat}{A data file as read in by \code{r4ss::SS_readdat}.}

\item{temp_ctl}{A subset of the control file representing the parameter section of interest (i.e. MG, SR, Q, or Selectivity)}

\item{base_range}{the difference between the base parameters max and min bounds}

\item{baseparm}{The value of the base parameter}

\item{base_bounds}{The min and max bounds of the base parameter}
}
\value{
A modified parameter value series that incorporates the appropriate time varying block effects.
}
\description{
Update a sequence of base parameter annual values to account for a time varying block effects
}
\author{
Nathan Vaughan
}
