% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_extend_OM.R
\name{update_basevals_dev}
\alias{update_basevals_dev}
\title{Update a sequence of base parameter annual values to account for a time varying deviation effects}
\usage{
update_basevals_dev(
  base_vals,
  temp_dev,
  dev_seq,
  current_par,
  temp_ctl,
  base_range,
  base_bounds
)
}
\arguments{
\item{base_vals}{A vector of base parameter values that will be updated to include the impact of a time varying deviations}

\item{temp_dev}{The time varying parameter lines for the deviations on the base parameter}

\item{dev_seq}{A vector of the parameter deviations to be applied to the base values}

\item{current_par}{The index of the current parameter being updated}

\item{temp_ctl}{A subset of the control file representing the parameter section of interest (i.e. MG, SR, Q, or Selectivity)}

\item{base_range}{the difference between the base parameters max and min bounds}

\item{base_bounds}{The min and max bounds of the base parameter}
}
\value{
A modified parameter series that incorporates the appropriate deviations.
}
\description{
Update a sequence of base parameter annual values to account for a time varying deviation effects
}
\author{
Nathan Vaughan
}
