% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_extend_OM.R
\name{update_basevals_env}
\alias{update_basevals_env}
\title{Update a sequence of base parameter annual values to account for a time varying environmental effects}
\usage{
update_basevals_env(
  base_vals,
  base_years,
  temp_env,
  current_par,
  timeseries,
  temp_ctl,
  dat,
  base_range,
  base_bounds,
  parlist
)
}
\arguments{
\item{base_vals}{A vector of base parameter values that will be updated to include the impact of a time varying environmental effects}

\item{base_years}{A vector of years for which the base values are needed}

\item{temp_env}{The time varying parameter lines for the environmental effects on the base parameter}

\item{current_par}{The index of the current parameter being updated}

\item{timeseries}{The timeseries table from \code{r4ss::SS_output()}.}

\item{temp_ctl}{A subset of the control file representing the parameter section of interest (i.e. MG, SR, Q, or Selectivity)}

\item{dat}{A datafile as read in by r4ss::SS_readdat}

\item{base_range}{the difference between the base parameters max and min bounds}

\item{base_bounds}{The min and max bounds of the base parameter}

\item{parlist}{A parameter file as read in by \code{r4ss::SS_readpar_3.30}}
}
\value{
A modified parameter series that incorporates the appropriate time varying environmental effects.
}
\description{
Update a sequence of base parameter annual values to account for a time varying environmental effects
}
\author{
Nathan Vaughan
}
