% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develtools.R
\name{update_ss3_version}
\alias{update_ss3_version}
\title{function for package developers to update SS3 input files}
\usage{
update_ss3_version(dir_models = "inst/extdata/models")
}
\arguments{
\item{dir_models}{directory containing the model input files}
}
\description{
function for package developers to update SS3 input files
}
