% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{add_sample_struct}
\alias{add_sample_struct}
\title{Add in years of sampling data needed}
\usage{
add_sample_struct(sample_struct, dat)
}
\arguments{
\item{sample_struct}{A optional list including which years, seasons, and fleets
should be  added from the OM into the EM for different types of data.
If NULL, the data structure will try to be inferred from the pattern found
for each of the datatypes within the EM datafiles. Include this structure
for the number of years to extend the model out. Note that the data should
be specified using the list component names and column names as in would be
used in \code{r4ss::SS_readdat()}. The \code{run_SSMSE_iter} function
examples give an example of what this structure should be. Running the
function \code{create_sample_struct()} will also produce a sample_struct
object in the correct form. Can be NULL only when MS is not EM.}

\item{dat}{A datafile as read in by r4ss::SS_readdat}
}
\description{
Add in years of sampling data needed
}
