% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvirEM.R, R/RatioBiasEM.R
\name{biasEM_change_dat}
\alias{biasEM_change_dat}
\title{Change dataset from OM into format for EM}
\usage{
biasEM_change_dat(
  OM_datfile,
  EM_datfile,
  EM_dir,
  do_checks = TRUE,
  verbose = FALSE,
  sample_struct_hist = NULL
)

biasEM_change_dat(
  OM_datfile,
  EM_datfile,
  EM_dir,
  do_checks = TRUE,
  verbose = FALSE,
  sample_struct_hist = NULL
)
}
\arguments{
\item{OM_datfile}{Filename of the datfile produced by the OM within the
EM_dir.}

\item{EM_datfile}{Filename of the datfile from the original EM within the
EM_dir.}

\item{EM_dir}{Absolute or relative path to the Estimation model directory.}

\item{do_checks}{Should checks on the data be performed? Defaults to TRUE.}

\item{verbose}{Want verbose output? Defaults to FALSE.}

\item{sample_struct_hist}{historical sample structure object}
}
\value{
the new EM data file. Side effect is saving over the OM_dat file in
EM_dir.

the new EM data file. Side effect is saving over the OM_dat file in
EM_dir.
}
\description{
Change dataset from OM into format for EM

Change dataset from OM into format for EM
}
\examples{
\dontrun{
# TODO: Add example
}
\dontrun{
# TODO: Add example
}
}
\author{
Kathryn Doering modified by Cassidy Peterson
}
