% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_error.R
\name{build_impl_error}
\alias{build_impl_error}
\title{Build a list of implementation error vectors for each scenario and iteration.}
\usage{
build_impl_error(
  yrs,
  nyrs_assess,
  n_impl_error_groups,
  scope,
  impl_error_pattern,
  impl_error_pars,
  n_scenarios,
  iter_list,
  seed
)
}
\arguments{
\item{yrs}{the number of years to simulate implementation error for.}

\item{nyrs_assess}{the number of years between assessments.}

\item{n_impl_error_groups}{the number of fleets times number of seasons}

\item{scope}{scope over which implementation errors will be randomized.}

\item{impl_error_pattern}{how to simulate implementation errors.}

\item{impl_error_pars}{implementation error simulation parameters dependent on chosen pattern.}

\item{n_scenarios}{The number of scenarios simulated}

\item{iter_list}{a list of iteration names.}

\item{seed}{a list prespecified random seed values to enable replication of simulated implementation error}
}
\value{
A list of scenarios with lists of interations in each with a vector of
 implementation errors for each simulation year-fleet-season combination.
}
\description{
Build a list of implementation error vectors for each scenario and iteration.
}
