% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdevs.R
\name{build_rec_devs}
\alias{build_rec_devs}
\title{Build an array of recruitment deviation vectors for every scenario and iteration}
\usage{
build_rec_devs(
  rec_dev_pattern = c("none", "rand", "AutoCorr_rand", "AutoCorr_Spec", "vector"),
  yrs,
  iter_vec,
  scope = c("2", "1", "3"),
  rec_dev_pars = NULL,
  stddev = NULL,
  rec_autoCorr = NULL,
  seed = NULL
)
}
\arguments{
\item{rec_dev_pattern}{how to simulate recruitment devations.}

\item{yrs}{the number of years to simulate recruitment deviations for for each
scenario. An integer vector with length equal to the number of scenarios.}

\item{iter_vec}{An integer vector (with length n_scenarios) of number of
iterations by scenario. length(iter_vec) should be the number of scenarios simulated.}

\item{scope}{scope over which recruitment devations will be randomized.}

\item{rec_dev_pars}{recruitment devation simulation parameters dependent on chosen pattern.}

\item{stddev}{the standard deviation of simulated recruitment deviations}

\item{rec_autoCorr}{a list of auto-correlation paramters and standard deviations}

\item{seed}{a list prespecified random seed values to enable replication of simulated rec devs}
}
\value{
A list of scenarios with lists of interations in each with a vector of
 rec devs for each simulation year.
}
\description{
Build an array of recruitment deviation vectors for every scenario and iteration
}
