% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM_create_devs_list.R
\name{calc_par_trend}
\alias{calc_par_trend}
\title{Calculate the parameter trend}
\usage{
calc_par_trend(
  val_info,
  val_line = c("mean", "sd", "cv", "ar_1_phi"),
  ref_parm_value,
  vals_df,
  parname,
  parlist,
  ctl,
  par_section,
  dat
)
}
\arguments{
\item{val_info}{The line in the input df containing info about the parameter.}

\item{val_line}{Which line in val info to use.}

\item{ref_parm_value}{This is the historic parameter that the end trend value.
Can be NA if the there is no line in val_info for the given parameter}

\item{vals_df}{The dataframe of the parameter values by year. Use to get
start val and last year}

\item{parname}{Name of the parameter with devs from the SS3 model.
will reference, if using a relative method.}

\item{parlist}{A parameter file as read in by \code{r4ss::SS_readpar_3.30}}

\item{ctl}{A control file as read in by \code{r4ss::SS_readctl}.}

\item{par_section}{Which parameter section should this variable be in?}

\item{dat}{A data file as read in by \code{r4ss::SS_readdat}.}
}
\value{
A vector of values with length ncol(vals_df), the number of future
years.
}
\description{
Calculate the parameter trend
}
\author{
Kathryn Doering
}
