% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_om.R
\name{check_future_om_list_str}
\alias{check_future_om_list_str}
\title{Check the general structure of a future OM list and standardize values}
\usage{
check_future_om_list_str(future_om_list)
}
\arguments{
\item{future_om_list}{An optional list of lists including changes that should
be made after the end year of the input model. Each first level list element
outlines 1 change to be made to the operating model. To see an example, try
running \code{\link{create_future_om_list}}. Defaults to NULL, which implies
that the model will be extended forward in time assuming the original model
structure.}
}
\value{
The future_om_list with implicit arguments made explicit
}
\description{
Checks that a future OM list is valid. If any values are implicit,
then add these values. Does not check against arguments in the scenario,
just the generic structure
}
