% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/future_om.R
\name{check_future_om_list_vals}
\alias{check_future_om_list_vals}
\title{Check structure of a future OM list against the scen_list}
\usage{
check_future_om_list_vals(future_om_list, scen_list)
}
\arguments{
\item{future_om_list}{The future_om_list object to check.}

\item{scen_list}{The list object of scenarios specifying inputs, typically 
passed to or created by run_SSMSE.}
}
\value{
The future_om_list with implicit arguments made explicit
}
\description{
Checks that a future OM list is valid when compared with the scen_list inputs
}
