% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkinput.R
\name{check_sample_struct}
\alias{check_sample_struct}
\title{Check sample_struct_list}
\usage{
check_sample_struct(
  sample_struct,
  valid_names = list(catch = c("Yr", "Seas", "FltSvy", "SE"), CPUE = c("Yr", "Seas",
    "FltSvy", "SE"), lencomp = c("Yr", "Seas", "FltSvy", "Sex", "Part", "Nsamp"), agecomp
    = c("Yr", "Seas", "FltSvy", "Sex", "Part", "Ageerr", "Lbin_lo", "Lbin_hi", "Nsamp"))
)
}
\arguments{
\item{sample_struct}{The list to check. Should be a list including which years and
fleets should be added from the OM into the EM for different types of data.}

\item{valid_names}{The list to compare sample_struct to.}
}
\description{
Check that list object sample_struct_list has the expected form, including the
correct names, correct column names (as in r4ss), and that all values in the
dataframes are integer or numeric. This does not check for if numeric or
interger values make sense given the model used.
}
\author{
Kathryn Doering
}
