% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_init_mod_files}
\alias{clean_init_mod_files}
\title{clean the initial model files}
\usage{
clean_init_mod_files(
  OM_out_dir,
  EM_out_dir = NULL,
  MS = "EM",
  overwrite = FALSE
)
}
\arguments{
\item{OM_out_dir}{The directory containing the OM direcatory}

\item{EM_out_dir}{The directory containing the EM directory. If an EM will
not be used, can be set to NULL}

\item{MS}{The management strategy}

\item{overwrite}{Should the data files be overwiritten?}
}
\description{
clean the initial model files
}
