% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_r4ss_names}
\alias{convert_to_r4ss_names}
\title{Convert user input to r4ss data names}
\usage{
convert_to_r4ss_names(
  sample_struct,
  convert_key = data.frame(df_name = c(rep("catch", 3), rep("CPUE", 3), rep("lencomp",
    5), rep("agecomp", 8)), r4ss_name = c("year", "seas", "fleet", "year", "seas",
    "index", "Yr", "Seas", "FltSvy", "Gender", "Part", "Yr", "Seas", "FltSvy", "Gender",
    "Part", "Ageerr", "Lbin_lo", "Lbin_hi"), sample_struct_name = c("Yr", "Seas",
    "FltSvy", "Yr", "Seas", "FltSvy", "Yr", "Seas", "FltSvy", "Sex", "Part", "Yr",
    "Seas", "FltSvy", "Sex", "Part", "Ageerr", "Lbin_lo", "Lbin_hi"))
)
}
\arguments{
\item{sample_struct}{The sampling structure specified by the user to a 
run_SSMSE* wrapper function}

\item{convert_key}{Data frame defining how r4ss names relate to the sample_struct
names. For now, a 1:1 relationship is assumed.}
}
\description{
Convert user input to r4ss data names
}
