% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_model_files}
\alias{copy_model_files}
\title{Copy OM and EM model files}
\usage{
copy_model_files(
  OM_in_dir,
  OM_out_dir,
  MS = "not_EM",
  EM_in_dir,
  EM_out_dir,
  verbose = FALSE
)
}
\arguments{
\item{OM_in_dir}{The OM input directory}

\item{OM_out_dir}{The OM output directory}

\item{MS}{The management strategy. EM_in_dir and EM_out_dir will only be
used if MS = "EM".}

\item{EM_in_dir}{The EM input directory. Only needed if MS = "EM".}

\item{EM_out_dir}{The EM output directory. Only needed if MS = "EM".}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\value{
TRUE, if copying is successful
}
\description{
Copy OM and EM model files from input to output location.
}
