% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_model_files}
\alias{copy_model_files}
\title{Copy OM and EM model files}
\usage{
copy_model_files(
  OM_in_dir = NULL,
  OM_out_dir = NULL,
  EM_in_dir = NULL,
  EM_out_dir = NULL,
  verbose = FALSE
)
}
\arguments{
\item{OM_in_dir}{Relative or absolute path to the operating model, if using a
model outside of the SSMSE package. Should be a string.}

\item{OM_out_dir}{The full path to the directory in which the OM is run.}

\item{EM_in_dir}{Relative or absolute path to the estimation model,}

\item{EM_out_dir}{Relative or absolute path to the estimation model, if using a
model outside of the SSMSE package.}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\value{
TRUE, if copying is successful
}
\description{
Copy OM and EM model files from input to output location.
}
