% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{create_OM}
\alias{create_OM}
\title{Create the OM by adding dummy data}
\usage{
create_OM(
  OM_out_dir,
  overwrite = TRUE,
  add_dummy_dat = FALSE,
  writedat = TRUE,
  verbose = FALSE,
  nyrs_assess = NULL,
  rec_devs = NULL,
  verify_OM = TRUE,
  seed = NULL
)
}
\arguments{
\item{OM_out_dir}{The full path to the directory in which the OM is run.}

\item{overwrite}{Overwrite existing files with matching names?}

\item{add_dummy_dat}{Should dummy data be added to indices and comps for each
year so that expected values and sampling is obtained for years other than
those that already have data? Defaults to FALSE.}

\item{writedat}{Should a new datafile be written? Only used if add_dummy_dat
is \code{TRUE}.}

\item{verbose}{Want verbose output? Defaults to FALSE.}

\item{nyrs_assess}{The number of years between assessments. This is used to
structure the forecast file for use in the OM.}

\item{rec_devs}{Vector of recruitment deviations for simulation.}

\item{verify_OM}{Should the model be run without estimation and some basic
checks done to verify that the OM can run? Defaults to TRUE.}

\item{seed}{input seed to allow reproducible SS results.}
}
\value{
A new datafile as read in for r4ss, but with dummy data added.
}
\description{
This function manipulates the OM as needed so that it can be used as an
 operating model.
}
\author{
Kathryn Doering & Nathan Vaughan
}
