% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{create_OM}
\alias{create_OM}
\title{Create the OM}
\usage{
create_OM(
  OM_out_dir,
  overwrite = TRUE,
  writedat = TRUE,
  verbose = FALSE,
  nyrs = NULL,
  nyrs_assess = NULL,
  nscen = 1,
  scen_name = NULL,
  niter = 1,
  rec_devs = NULL,
  future_om_list = NULL,
  verify_OM = TRUE,
  sample_struct_hist = NULL,
  seed = NULL
)
}
\arguments{
\item{OM_out_dir}{The full path to the directory in which the OM is run.}

\item{overwrite}{Overwrite existing files with matching names?}

\item{writedat}{Should a new datafile be written?}

\item{verbose}{Want verbose output? Defaults to FALSE.}

\item{nyrs}{Number of years beyond the years included in the OM to run the
MSE. A single integer value.}

\item{nyrs_assess}{The number of years between assessments. This is used to
structure the forecast file for use in the OM.
@param nscen The scenario number
@param scen_name The scenario name
@param niter the iteration number}

\item{rec_devs}{Vector of recruitment deviations for simulation.}

\item{future_om_list}{An optional list of lists including changes that should
be made after the end year of the input model. Each first level list element
outlines 1 change to be made to the operating model. This will identify which
parameters to turn deviations on for}

\item{verify_OM}{Should the model be run without estimation and some basic
checks done to verify that the OM can run? Defaults to TRUE.}

\item{sample_struct_hist}{The historical sample structure, if specified by 
the user. Defaults to NULL, which means to use the same sample structure as
the historical data.}

\item{seed}{input seed to allow reproducible SS results.}
}
\value{
A modified datafile
}
\description{
This function manipulates the OM as needed so that it can be used as an
 operating model.
}
\author{
Kathryn Doering & Nathan Vaughan
}
