% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{create_OM}
\alias{create_OM}
\title{Create the OM from a stock assessment model}
\usage{
create_OM(OM_dir, SA_dir, overwrite = FALSE, add_dummy_dat = FALSE,
  verbose = FALSE)
}
\arguments{
\item{OM_dir}{The full path to the OM directory}

\item{SA_dir}{The full path to the EM directory}

\item{overwrite}{Overwrite existing files with matching names?}

\item{add_dummy_dat}{Add dummy data to indices and comps for each year so 
that expected values and sampling is obtained for years other than those 
that already have data? Defaults to FALSE.}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\value{
OM_dir, because this function is used mainly for its side effects.
}
\description{
This function copies over the stock assessment model and manipulates it as
needed so that it can be used as an operating model.
}
\author{
Kathryn Doering
}
