% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop_OMs.R
\name{develop_OMs}
\alias{develop_OMs}
\title{Develop different operating models}
\usage{
develop_OMs(
  OM_dir,
  out_dir = getwd(),
  par_name,
  par_vals,
  refit_OMs = TRUE,
  hess = TRUE
)
}
\arguments{
\item{OM_dir}{Path to the original operating model}

\item{out_dir}{Path where the new models will be written. Defaults to the
current working directory.}

\item{par_name}{Name of the parameter to modify}

\item{par_vals}{Vector of parameter values to modify in the OM. Assume these
will be fixed so phase will be set as negative.}

\item{refit_OMs}{Should the models be refit to data? Defaults to TRUE}
}
\description{
THis is a utility to help a user create new operating models starting from the
same model. For now, it is only possible to adjust 1 parameter value
}
