% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendOM.R
\name{extend_OM}
\alias{extend_OM}
\title{Extend the OM forward using next years' catch}
\usage{
extend_OM(catch, OM_dir, dummy_dat_scheme = NULL, nyrs = 3,
  verbose = FALSE)
}
\arguments{
\item{catch}{A dataframe of catch values and its associated information to
 add to the OM. The column names are the same as in an SS data file (e.g., 
 year,    season, fleet,    catch,    catch_se).
length of the number of years (only works when catch is for 1 fleet)}

\item{OM_dir}{The full path to the OM directory.}

\item{dummy_dat_scheme}{The sampling scheme for dummy data. A list of lists}

\item{nyrs}{An integer value of years to extend the model forward. Defaults
to an arbitrary value of 3.}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\description{
Add in the catch values for the next years, extend the OM forward for the 
appropriate number of years.
}
