% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendOM.R
\name{extend_OM}
\alias{extend_OM}
\title{Extend the OM forward using next years' catch}
\usage{
extend_OM(
  catch,
  discards,
  OM_dir,
  sample_struct = NULL,
  nyrs_extend = 3,
  write_dat = TRUE,
  rec_devs = NULL,
  impl_error = NULL,
  verbose = FALSE
)
}
\arguments{
\item{catch}{A dataframe of catch values and its associated information to
 add to the OM. The column names are the same as in an SS data file (e.g.,
 year,    season, fleet,    catch,    catch_se).
length of the number of years (only works when catch is for 1 fleet)}

\item{discards}{A dataframe of discard values an dassociated information to
add to the OM. The column names are as in an SS datafile.}

\item{OM_dir}{The full path to the OM directory.}

\item{sample_struct}{The sample structure dataframe.}

\item{nyrs_extend}{An integer value of years to extend the model forward. Defaults
to an arbitrary value of 3.}

\item{write_dat}{Should the datafile be overwritten? Defaults to TRUE.}

\item{rec_devs}{The recruitment deviations}

\item{impl_error}{The implementation error}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\value{
A new dat list object (format as created by r4ss::SS_readdat) that
 has been extended forward  as if read in by r4ss function SS_readdat
}
\description{
Add in the catch values for the next years, extend the OM forward for the
appropriate number of years.
}
\author{
Kathryn Doering & Nathan Vaughan
}
