% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_F.R
\name{get_F}
\alias{get_F}
\title{Get the Fishing mortality from the timeseries Report.sso table}
\usage{
get_F(timeseries)
}
\arguments{
\item{timeseries}{from SSoutput}
}
\value{
a data frame with Fs by Area, Yr, Era, Seas, and Fleet (long format)
}
\description{
Get the Fishing mortality from the timeseries Report.sso table
}
