% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_F.R
\name{get_F}
\alias{get_F}
\title{get the Fishing mortality from a forecasting model run}
\usage{
get_F(timeseries, units_of_catch, Nfleet)
}
\arguments{
\item{timeseries}{from SSoutput}

\item{units_of_catch}{From datalist}

\item{Nfleet}{number of fishing fleets (does not include surveys)}
}
\value{
a data frame of Fs and tetained catch by units (numbers or biomass) 
 seems to have some issues, though
}
\description{
get the Fishing mortality from a forecasting model run
}
