% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_F.R
\name{get_F}
\alias{get_F}
\title{Get the Fishing mortality from the timeseries Report.sso table}
\usage{
get_F(timeseries, fleetnames, fleetnames_all)
}
\arguments{
\item{timeseries}{The timeseries table from \code{r4ss::SS_output()}.}

\item{fleetnames}{A vector of fleet names, in the order they appear in the
SS3 model.}

\item{fleetnames_all}{A vector of ALL fleet names that are in the model in
the order that they are specified in the model. This vector helps the
function know which order the fleets appear in the model.}
}
\value{
a list containing: F_df, a long dataframe with F by Yr, Era, Seas,
and fleet; F_rate, a data frame with F for the time frame of the model
only by Yr, Seas, and Fleet, ordered as the ss.par file expects; init_F,
a named vector of initial F values by Season and Fleet, ordered (and named)
as SS3 expects; and F_rate_fcast, a dataframe of forecasted F by Yr, Seas,
and fleet, ordered as SS3 would expect in F_rate.
}
\description{
Get the Fishing mortality from the timeseries Report.sso table
}
