% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_struct.R
\name{get_full_sample_struct}
\alias{get_full_sample_struct}
\title{Get the full sample structure from user input}
\usage{
get_full_sample_struct(sample_struct, OM_out_dir)
}
\arguments{
\item{sample_struct}{The sample structure, as defined by the user. This need
not define all the sampling structure if it can be unambiguously determined
from the OM data file.}

\item{OM_out_dir}{The directory containing the operating model files}
}
\value{
A list of the full sample structure, using names as input by the user
input by the user (not r4ss names).
}
\description{
Get the ful sample structure from user input by looking at the OM data. If it
cannot be unambigously determined, this function will return an error
describing what additional user input is required.
}
