% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impl_error.R
\name{get_impl_error_matrix}
\alias{get_impl_error_matrix}
\title{Put implementation error of 0 into a matrix}
\usage{
get_impl_error_matrix(yrs)
}
\arguments{
\item{yrs}{a vector of years}
}
\value{
A length(yrs) row, 2 column matrix containing the years in the first
column and the implementation error values in the second.
}
\description{
Put implementation error of 0 into a matrix
}
