% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MS.R
\name{get_no_EM_catch_df}
\alias{get_no_EM_catch_df}
\title{Get the data frame of catch for the next iterations when not using an
estimation model.}
\usage{
get_no_EM_catch_df(OM_dir, yrs, MS = "last_yr_catch")
}
\arguments{
\item{OM_dir}{The OM directory.}

\item{yrs}{A vector of years}

\item{MS}{Can be either "no_catch" or "last_yr_catch"}
}
\value{
A dataframe of future catch.
}
\description{
Get the data frame of catch for the next iterations when not using an
estimation model.
}
