% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdevs.R
\name{get_rec_devs_matrix}
\alias{get_rec_devs_matrix}
\title{Put recdevs into a matrix}
\usage{
get_rec_devs_matrix(yrs, rec_devs)
}
\arguments{
\item{yrs}{a vector of years}

\item{rec_devs}{a vector of recdevs}
}
\value{
A length(yrs) row, 2 column matrix containing the years in the first
column and the rec_devs values in the second.
}
\description{
Put recdevs into a matrix
}
