% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_in_dirs}
\alias{locate_in_dirs}
\title{create the OM directory}
\usage{
locate_in_dirs(OM_name, OM_in_dir)
}
\arguments{
\item{OM_name}{Name of OM model.}

\item{OM_in_dir}{Relative or absolute path to the operating model. NULL if 
using SSMSE package model.}

\item{out_dir}{The directory to which to write output. IF NULL, will default
to the working directory.}

\item{niter}{The number iteration}

\item{MS}{The management strategy. If "EM", will create the EM folders.
Otherwise, can be any name, defaulting to "not_EM".}

\item{EM_name}{Name of the EM model. NULL if MS != "EM" (default value)}

\item{EM_in_dir}{Relative or absolute path to the estimation model. NULL if 
using SSMSE package model or MS != "EM" (default value)}
}
\value{
A list with 2 named components each of length 1 characters. The
 components are: OM_dir, where OM will be run, and OM_in_dir, where the model
 files will be copied from.
}
\description{
Create an OM directory within the out_dir specified (named by the value of
niter)
}
