% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{locate_in_dirs}
\alias{locate_in_dirs}
\title{Locate the OM model files}
\usage{
locate_in_dirs(OM_name = NULL, OM_in_dir = NULL)
}
\arguments{
\item{OM_name}{Name of the operating model (OM).}

\item{OM_in_dir}{Relative or absolute path to the operating model, if using a
model outside of the SSMSE package. Should be a string.}
}
\value{
A list with a component, OM_in_dir, which contains the model location
}
\description{
Locate the OM model files
}
