% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_MS.R
\name{parse_MS}
\alias{parse_MS}
\title{Parse management strategy options}
\usage{
parse_MS(
  MS,
  EM_out_dir = NULL,
  EM_init_dir = NULL,
  init_loop = TRUE,
  OM_dat,
  OM_out_dir = NULL,
  verbose = FALSE,
  nyrs_assess,
  dat_yrs,
  sample_struct = NULL,
  interim_struct = NULL,
  seed = NULL
)
}
\arguments{
\item{MS}{The management strategy to use. Current options
are:  \code{"last_yr_catch"} which uses the previous year's catch; 
\code{"no_catch"} which uses 0 catch; \code{"EM"} which uses an stock 
synthesis model as the estimation method and the management strategy as 
defined in the forecast file of the stock synthesis estimation method;
\code{"Interim"} to modify catch based on survey predictions between
assessments. Users can also specify their own managment strategies as an
function. To use the function, it must be available in the global
enviroment and specified by name in MS. For example, if the function is
called "my_ms" then the user should make it available in the global 
environment and specify MS = "my_ms".}

\item{EM_out_dir}{Relative or absolute path to the estimation model, if using a
model outside of the SSMSE package. Note that this value should be NULL if
\code{MS} has a value other than \code{"EM"}.}

\item{EM_init_dir}{Initialization director that retains the reference files for interim assessments}

\item{init_loop}{Logical. If this is the first initialization loop of the
MSE, \code{init_loop} should be TRUE. If it is in further loops, it should
be FALSE.}

\item{OM_dat}{An valid SS data file read in using r4ss. In particular,
this should be sampled data.}

\item{OM_out_dir}{The full path to the directory in which the OM is run.}

\item{verbose}{Want verbose output? Defaults to FALSE.}

\item{nyrs_assess}{The number of years between assessments. E.g., if an
assessment is conducted every 3 years, put 3 here. A single integer value.}

\item{dat_yrs}{Which years should be added to the new model? Ignored if
init_loop is TRUE.}

\item{sample_struct}{An optional list including which years and fleets should be
added from the OM into the EM for different types of data. If NULL, the data
structure will try to be infered from the pattern found for each of the
datatypes within EM_datfile. Ignored if init_loop is TRUE.}

\item{interim_struct}{An optional including how many years to average over,
fleet weights, the scaling rate (Beta) of catch relative to the index change for each fleet,
and the reference year for each fleet (either a fixed year or <=0 relative to end_yr, fixed year
will stay constant during simulation while relative year will progress with simulation).}

\item{seed}{a random seed to initialize SS runs}
}
\description{
This function matches each management strategy with its correct method. And
checks for errors.
}
\author{
Kathryn Doering & Nathan Vaughan
}
