% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{rm_sample_struct_hist}
\alias{rm_sample_struct_hist}
\title{Remove the historical sampling structure}
\usage{
rm_sample_struct_hist(sample_struct_hist, dat)
}
\arguments{
\item{sample_struct_hist}{The historical sampling structure as specified by
the user. Can be NULL, in which case no values will be removed. Uses the
names as in the r4ss dataframes.}

\item{dat}{The data file, as read in using r4ss}
}
\description{
Remove the historical sampling structure
}
