% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate_EM.R
\name{run_EM}
\alias{run_EM}
\title{Run the estimation model}
\usage{
run_EM(
  EM_dir,
  hess = FALSE,
  check_converged = TRUE,
  set_use_par = FALSE,
  seas = 1,
  catch_se = "most_common_value",
  verbose = FALSE
)
}
\arguments{
\item{EM_dir}{Absolute or relative path to the estimation model directory}

\item{hess}{Get the hessian during model run? Defaults to FALSE. Not
estimating the hessian will speed up the run, but no estimates of error will
be generated.}

\item{check_converged}{Perform checks to see if the model converged? Defaults
to TRUE.}

\item{set_use_par}{Should input values be read from the .par file? If TRUE,
will change setting in the starter file; otherwise, will use the setting
already in the starter file, which may or may not read from the .par file.}

\item{seas}{Season for catch to be output. Defaults to 1.}

\item{catch_se}{Catch standard error for catch to be output. Either a single
numeric value (e.g., 0.1), a vector the same length as \code{nyrs_proj}, or
a string specifying how to get catch_se from the data file (options are: 
"most_common_value" which uses the most common value in the data as 
\code{catch_se} or "only_value" which will use the only value provided in 
the data as \code{catch_se}; if there are multiple values for standard error, 
a fatal error will occur). Default is "most_common_value"}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\value{
The future catch in a dataframe, based on the EM run and the harvest
strategy outlined within the SS forecasting module.
}
\description{
Runs the estimation model and performs checks if desired.
}
\author{
Kathryn Doering
}
