% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initOM.R
\name{run_OM}
\alias{run_OM}
\title{Initial run of the OM}
\usage{
run_OM(OM_dir, boot = TRUE, nboot = 1, init_run = FALSE, verbose = FALSE)
}
\arguments{
\item{OM_dir}{The full path to the OM directory}

\item{boot}{Return the bootstrap dataset? If TRUE, function returns the 
number bootstrapped dataset specified in \code{nboot}. If FALSE, it returns
the expected values.}

\item{nboot}{The number bootstrapped data set. This value is only used if 
\code{boot = TRUE}. Note that this numbering does NOT correspond with the
numbering in section of r4ss::SS_readdat. E.g., specifying section = 3 in 
SS_readdat is equivalent to specifying nboot = 1.}

\item{init_run}{Is this the initial iteration of the OM? Defaults to FALSE.}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\description{
This function is used to initialize the OM and get either expected values 
or bootstrap.
}
\author{
Kathryn Doering
}
