% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSSMSE.R
\name{run_SSMSE}
\alias{run_SSMSE}
\title{run an MSE using SS OMs}
\usage{
run_SSMSE()
}
\description{
High level function to run a management strategy evaluation using Stock
Synthesis as the Operating model(s)
}
