% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSSMSE.R
\name{run_SSMSE_scen}
\alias{run_SSMSE_scen}
\title{Run an MSE scenario using SS OM}
\usage{
run_SSMSE_scen()
}
\description{
High level function to run 1 scenario (but potentially many iterations) for 
a management strategy evaluation using Stock Synthesis as the Operating Model
}
