% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSSMSE.R
\name{run_SSMSE_scen}
\alias{run_SSMSE_scen}
\title{Run an MSE scenario using SS OM}
\usage{
run_SSMSE_scen(
  scen_name = "scen_1",
  out_dir_scen = NULL,
  iter = 1:2,
  OM_name = "cod",
  OM_in_dir = NULL,
  EM_name = NULL,
  EM_in_dir = NULL,
  MS = "no_catch",
  use_SS_boot = TRUE,
  nyrs = 100,
  nyrs_assess = 3,
  rec_devs_scen = NULL,
  impl_error = NULL,
  sample_struct = NULL,
  verbose = FALSE
)
}
\arguments{
\item{scen_name}{Name of the scenario. The directory containing all the model
runs the scenario will be stored within a folder of this name.}

\item{out_dir_scen}{The directory to which to write output. IF NULL, will
default to the working directory.}

\item{iter}{A vector of integers to refer each iteration of the scenario.
The length of this vector will be the number of iterations run for the
scenario.}

\item{OM_name}{Name of a valid Stock Synthesis stock assessment model from
which to create the OM. Currently, only allows models in the package, so
valid inputs are: \code{"cod"}.}

\item{OM_in_dir}{Relative or absolute path to the operating model, if using a
model outside of the SSMSE package.}

\item{EM_name}{Should be NULL unless \code{MS = "EM"}. Name of a valid Stock
Synthesis stock assessment model to use as an EM. If the value of EM_name
is NULL and  \code{MS = "EM"}, then SSMSE will look for the estimation model
in the path specified in EM_in_dir. valid inputs for EM_name are: \code{"cod"}
or \code{NULL}.}

\item{EM_in_dir}{Relative or absolute path to the estimation model, if using a
model outside of the SSMSE package. Note that this value should be NULL if
\code{MS} has a value other than "EM".}

\item{MS}{The management strategy to use. Current options
are:  \code{"last_yr_catch"} which uses the previous year's catch; 
\code{"no_catch"} which uses 0 catch; and \code{"EM"} which uses the
management strategy in the estimation model.}

\item{use_SS_boot}{Should a bootstrapped data set generated by SS be used?
Defaults to TRUE.}

\item{nyrs}{Number of years beyond the years included in the OM to run the
MSE. A single integer value.}

\item{nyrs_assess}{The number of years between assessments. E.g., if an
assessment is conducted every 3 years, put 3 here. A single integer value.
(NOTE: we could make this more flexible by instead reading in a vector of
assessment years, so users could specify irregular numbers of yrs between
assessments.)}

\item{rec_devs_scen}{List containing a recruitment deviation vector for each
iteration.}

\item{impl_error}{List containing an implementation error vector for each
iteration.}

\item{sample_struct}{A optional list including which years, seasons, and fleets
should be  added from the OM into the EM for different types of data.
If NULL, the data structure will try to be infered from the pattern found
for each of the datatypes within the EM datafiles. Include this strucutre
for the number of years to extend the model out.}

\item{verbose}{Want verbose output? Defaults to FALSE.}
}
\description{
High level function to run 1 scenario (but potentially many iterations) for
a management strategy evaluation using Stock Synthesis as the Operating Model
}
\examples{
\dontrun{
# Create a temporary folder for the output and set the working directory:
  temp_path <- file.path(tempdir(), "run_SSMSE_scen-example")
  dir.create(temp_path, showWarnings = FALSE)

  # run 2 iteration and 1 scenario of SSMSE
  run_SSMSE_scen(scen_name = "no_catch",
                 iter = 1:2,
                 OM_name = "cod",
                 MS = "no_catch",
                 out_dir_scen = temp_path,
                 nyrs = 6,
                 nyrs_assess = 3
                 )
  unlink(temp_path, recursive = TRUE)
}

}
\author{
Kathryn Doering & Nathan Vaughan
}
