% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_MSE_seeds}
\alias{set_MSE_seeds}
\title{Set the initial global, scenario, and iteration seeds}
\usage{
set_MSE_seeds(seed, scen_name_vec, iter_list)
}
\arguments{
\item{seed}{reads in the user specified seeds if any to allow replication of runs}

\item{scen_name_vec}{reads the user input scen_name_vec to determine the number of scenarios}

\item{iter_list}{reads the user specified iteration list to determine the number of 
iterations per scenario}
}
\description{
Set the initial global, scenario, and iteration seeds
}
