% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSMSE.R
\docType{package}
\name{SSMSE}
\alias{SSMSE-package}
\alias{SSMSE}
\title{SSMSE: A package for Management Strategy Evaluation (MSE) using Stock
Synthesis (SS)}
\description{
SSMSE is an R package for performing Management Strategy Evaluation
(MSE) using Stock Synthesis (SS). SS is used as the Operating Model (OM)
and, if the user desires, the Estimation model (EM). SSMSE allows existing
SS models to be used as the basis for an OM. These OMs are used in the MSE
framework provided by SSMSE to evaluate the implications of management
actions on a population given uncertainty.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nmfs-fish-tools/SSMSE}
  \item Report bugs at \url{https://github.com/nmfs-fish-tools/SSMSE/issues}
}

}
\author{
\strong{Maintainer}: Kathryn Doering \email{kathryn.doering@noaa.gov}

Authors:
\itemize{
  \item Nathan Vaughan \email{nathan.vaughan@noaa.gov}
}

}
