% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{SSMSE_summary_all}
\alias{SSMSE_summary_all}
\title{Get results in a list for 1 scenario}
\usage{
SSMSE_summary_all(
  dir = getwd(),
  scenarios = NULL,
  run_parallel = FALSE,
  n_cores = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dir}{Path to the directory containing the scenarios, either relative 
or absolute. Defaults to the working directory.}

\item{scenarios}{A character vector of scenarios in dir from which to extract
summaries. If left as NULL, the summaries will be extracted from all folders
in dir.}

\item{run_parallel}{Option to use parallel processing. Defaults to FALSE.
Note that running in parallel will only be faster if most of the scenarios
do not yet have scenario-level summary files .}

\item{n_cores}{How many cores to use if running in parallel. If is NULL, 
defaults to n_cores available - 1 (also capped at one less than the number
of cores available - 1)}

\item{overwrite}{Should existing summary files be overwritten if present}
}
\value{
A list of 3 data frames called scalar, ts, and
 dq (for derived quantities). These lists contain information for
 multiple model runs (estimation models and operating models) for 1
 iteration.Also writes 3 .csv files with the contents of this list of
 dataframes to dir and 3.csv files with scenario specific results in each of
 the scenario foldurs..
}
\description{
Get results in a list for 1 iteration, using ss3sim::get_results_iter
}
\seealso{
\code{\link[ss3sim]{get_results_all}}
}
